% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\docType{data}
\name{audio_time_per_frame}
\alias{audio_time_per_frame}
\title{Time to play a single frame of audio at 44100 kHz}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
audio_time_per_frame
}
\description{
Time to play a single frame of audio at 44100 kHz
}
\keyword{datasets}
