% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\name{init_mpeg}
\alias{init_mpeg}
\title{Initialise mpeg playback}
\usage{
init_mpeg(file, video = TRUE, audio = TRUE, audio_stream = 0)
}
\arguments{
\item{file}{mpeg1 filename}

\item{video}{logical. Should video be decoded?  Default: TRUE}

\item{audio}{logical. Should audio be decoded?  Default: TRUE}

\item{audio_stream}{audio stream to play. Default: 0}
}
\value{
mpeg decoding context
}
\description{
Initialise mpeg playback
}
