% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\name{mpeg_decode_video}
\alias{mpeg_decode_video}
\title{Decode a frame of video}
\usage{
mpeg_decode_video(ctx)
}
\arguments{
\item{ctx}{mpeg context (as created by \code{init_mpeg()})}
}
\value{
native raster containing the image data. This data is "owned" by 
        the decoding process and will be overwritten with each
        decoded frame
}
\description{
Decode a frame of video
}
