% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\name{mpeg_info}
\alias{mpeg_info}
\title{Information about mpeg stream}
\usage{
mpeg_info(ctx)
}
\arguments{
\item{ctx}{mpeg context (as created by \code{init_mpeg()})}
}
\value{
named list of information
}
\description{
Information about mpeg stream
}
