% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\name{mpeg_rewind}
\alias{mpeg_rewind}
\title{Rewind to beginning of mpeg}
\usage{
mpeg_rewind(ctx)
}
\arguments{
\item{ctx}{mpeg context (as created by \code{init_mpeg()})}
}
\value{
None.
}
\description{
Rewind to beginning of mpeg
}
