% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\name{mpeg_seek}
\alias{mpeg_seek}
\title{Seek to time within mpeg data}
\usage{
mpeg_seek(ctx, time, exact = FALSE)
}
\arguments{
\item{ctx}{mpeg context (as created by \code{init_mpeg()})}

\item{time}{time (seconds)}

\item{exact}{logical.  perform slow exact seek? Default: FALSE}
}
\value{
logical value indicating if seek was successful
}
\description{
Seek to time within mpeg data
}
