% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\docType{data}
\name{samples_per_channel_per_frame}
\alias{samples_per_channel_per_frame}
\title{Samples per channel per frame}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
samples_per_channel_per_frame
}
\description{
Samples per channel per frame
}
\keyword{datasets}
