% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeg.R
\name{mpeg_decode_audio}
\alias{mpeg_decode_audio}
\title{Decode a single frame of audio (1152 stereo samples)}
\usage{
mpeg_decode_audio(ctx)
}
\arguments{
\item{ctx}{mpeg context (as created by \code{init_mpeg()})}
}
\value{
numeric vector with 2 * 1152 floating point sample values. Each 
        sample is in the range [-1, 1]. The stereo channels are interleaved.
}
\description{
Decode a single frame of audio (1152 stereo samples)
}
